refactor(examples-tutorial-basis): relocate server_fn modules into their owning apps
Move `src/server_fn/polls.rs` → `src/apps/polls/server_fn.rs` and
`src/server_fn/users.rs` → `src/apps/users/server_fn.rs`, then drop the
now-empty `src/server_fn.rs` module declaration from the crate root. Each
app now keeps its server functions next to its models / views / urls /
client components, which:

- Removes the cross-cutting parallel hierarchy that had to mirror the
  app list (every app already declares its own `models`, `urls`,
  `client` etc. — `server_fn` was the odd one out).
- Makes adding a new app a single-directory operation: drop a new
  module under `apps/<app>/`, add it to `installed_apps!`, done. No
  edit to a sibling `server_fn/` tree.
- Matches the per-app structure used by `examples-twitter`
  (`src/apps/<app>/shared/server_fn.rs`), making future cross-example
  refactors easier.

The relocation itself is purely structural — server-function bodies,
DI parameters, and behaviour are unchanged. Imports are updated
mechanically (`crate::server_fn::<app>::*` →
`crate::apps::<app>::server_fn::*`) across `config/urls.rs`,
`client/components/{polls,users,nav}.rs`, and the doc comments on
`apps/<app>/models.rs`.

Verified with `cargo check` (native + wasm32) and the existing
sign-up + voting flows in the dev server.
