+++ examples/examples-tutorial-basis/src/apps/users.rs
@@ -2,9 +2,11 @@
 //!
 //! Provides session-based authentication for the tutorial-basis example.
 //! Defines a minimal `User` model and exposes server functions for login,
-//! logout, and current-user introspection via `crate::server_fn::users`.
+//! logout, sign-up, and current-user introspection via
+//! `crate::apps::users::server_fn`.
 
 #[cfg(native)]
 pub mod models;
 
+pub mod server_fn;
 pub mod urls;
