--- examples/examples-tutorial-basis/src/server_fn.rs
@@ -1,7 +1,0 @@
-//! Server functions
-//!
-//! This module defines server functions that can be called from the client.
-//! Server functions are compiled for both WASM (client stubs) and server (actual implementation).
-
-pub mod polls;
-pub mod users;
